[refactor]: 패키지 분리; controller, domain, view
