+++ src/main/java/racingcar/domain/collection/CarGroupConst.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.collection;
 
 public class CarGroupConst {
     public static final String NAME_SPLIT_REGEX = ",";
