+++ src/main/java/racingcar/domain/collection/History.java
@@ -1,9 +1,19 @@
-package racingcar;
+package racingcar.domain.collection;
+
+import racingcar.domain.model.Memento;
 
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 
+/**
+ * NOTE: Memento 의 일급 컬렉션
+ * FIXME: racingcar.domain.collection 가 적절한지 확인 필요.
+ * 자바의 Collection 데이터 타입에 종속되는 패키지명이므로 혼동을 줄 수 있음.
+ * Memento 클래스 한개만 의존하고 있으므로,
+ * composition 이나, aggregation 이라는 표현은 부적절해보임.
+ * FIXME: collection 보다 적절한 패키지명을 추천받을 필요 있음.
+ */
 public class History {
     private List<Memento> mementoList;
