+++ src/main/java/racingcar/domain/model/Car.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain.model;
+
+import racingcar.domain.strategy.MoveStrategy;
 
 public class Car {
     private int position;
