+++ src/main/java/racingcar/domain/model/Memento.java
@@ -1,9 +1,15 @@
-package racingcar;
+package racingcar.domain.model;
 
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 
+/**
+ * NOTE: Memento(세이브)의 단위가 List<Car>가 아니라,
+ * Serialized 된 csv 혹은 json String 이 될 수도 있으므로,
+ * Memento 는 Car 의 컬렉션이 아니라,
+ * 별도의 model 로 분류되는 게 합당하다.
+ */
 public class Memento {
     private List<Car> cars;
