+++ src/main/java/racingcar/domain/strategy/RandomStrategyConst.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.strategy;
 
 public class RandomStrategyConst {
