+++ src/main/java/racingcar/view/InputViewConst.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 public class InputViewConst {
     public final static String ASK_CARS = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
