+++ src/main/java/racingcar/view/OutputViewConst.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 public class OutputViewConst {
     public final static String EXECUTION_RESULT_MSG = "실행 결과";
