+++ src/test/java/racingcar/RacingGameTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.aggregate.RacingGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
